Issue #9 Musician channel not joining if the player hasn't joined any channel
Fixed MusicianComm joining system to make it more robust and handle edge cases.
Requests are no longer blocked if sent in group channel or whisper when the communication channel is not ready yet.
Added a message when joined communication channel but alone ("There is no other music fan around yet.")